Adicionar base para o mcmc manchado
